fix: review card 스토리북 오류 수정
